requirements.txt
# sql support
psycopg2
SQLAlchemy

# firebase rest support: https://github.com/ozgur/python-firebase
requests
python-firebase
